refactor: Move  error to  package
