+++ src/languages/domain/errors/languages_errors.go
@@ -11,16 +11,3 @@ func (err *LangNotFoundError) Error() string {
 func (err *LangNotFoundError) StatusCode() int {
 	return http.StatusNotFound
 }
-
-type StaticFilesMicroserviceError struct {
-	Code    int
-	Message string
-}
-
-func (err *StaticFilesMicroserviceError) Error() string {
-	return err.Message
-}
-
-func (err *StaticFilesMicroserviceError) StatusCode() int {
-	return err.Code
-}
